move content.ts => content.json and update references
