+++ src/types.ts
@@ -1,4 +1,4 @@
-import type topics from "./data/content";
+import type topics from "./data/content.json";
 
 export const GeoTypes = ["ew", "lad", "msoa", "oa"];
 export type GeoType = typeof GeoTypes[number];
